mouse detection fix
NOTE - This needs to be looked at again. The mouse coordinates in the detection are not accurate, and I don't know why. It is most likely a sheet issue not buswire/symbol as the ports/wires seem to be correctly positioned. Sheet thinks that they are slightly off those positions, so I have translated the mouse coordinates to account for this but this needs properly looking into later.