fix SheetBeautifyHelpers, find almost-straight-wires
Fixed DevMode bug where wire tables would collapse upon clicking the table. This should only happen when clicking the summary tag.

Made choosing beautify model menu collapsible.

Updated sheet stats with algorithm to find near-straight wires.

Tested SheetBeautifyHelpers and noticed that removeSingleWireInvisibleSegments will get rid of nubs (ruins draggability functionality). Added a warning to the helper so it's only used for heuristics that do not modify the sheet, and if it must be used, I provided another helper to restore the nubs

Also created a variant of removeSingleWireInvisibleSegments to only act on a single wire, instead of a BusWireT.Model.Wires . See removeSingleWireInvisibleSegments.